chore(Typings): improving and simplifying typings to match latest... need help with Pipes.. if anyone can help here, please let me know.
