+++ src/app/bootstrap.ts
@@ -26,8 +26,7 @@ import {
 /*
  * Angular Modules
  */
-import {httpInjectables} from 'angular2/http';
-import {formInjectables} from 'angular2/forms';
+import {httpInjectables, formInjectables} from 'angular2/angular2';
 import {routerInjectables} from 'angular2/router';
 
 /*
