+++ src/app/components/dashboard.ts
@@ -3,8 +3,7 @@
 /*
  * Angular 2
  */
-import {Component, View, Directive} from 'angular2/annotations';
-import {ElementRef} from 'angular2/core';
+import {Component, View, Directive, ElementRef} from 'angular2/angular2';
 
 /*
  * TODO: refactor
