+++ src/app/components/rxjs-examples/autosuggest/Github.ts
@@ -1,7 +1,6 @@
 /// <reference path="../../../../typings/_custom.d.ts" />
 
-import {bind, Injectable} from 'angular2/di'
-import {Http} from 'angular2/http';
+import {bind, Injectable, Http} from 'angular2/angular2'
 
 import {ISearchable} from 'ISearchable';
