+++ src/app/components/rxjs-examples/autosuggest/Search.ts
@@ -1,8 +1,7 @@
 /// <reference path="../../../../typings/_custom.d.ts" />
 
 // Angular 2
-import {Component, View} from 'angular2/angular2';
-import {NgFor} from 'angular2/directives';
+import {Component, View, NgFor} from 'angular2/angular2';
 
 import {Autosuggest} from './Autosuggest';
