+++ src/app/components/rxjs-examples/rxjs-examples.ts
@@ -1,8 +1,7 @@
 /// <reference path="../../../typings/_custom.d.ts" />
 
 // Angular 2
-import {Component, View} from 'angular2/angular2';
-import {NgIf, CSSClass} from 'angular2/directives';
+import {Component, View, NgIf, CSSClass} from 'angular2/angular2';
 
 import {Search} from './autosuggest/Search';
 import {Timeflies} from './timeflies/timeflies';
