+++ src/app/directives/Autofocus.ts
@@ -1,6 +1,5 @@
 /// <reference path="../../typings/_custom.d.ts" />
-import {Directive} from 'angular2/annotations';
-import {ElementRef} from 'angular2/core';
+import {Directive, ElementRef} from 'angular2/angular2';
 // Simple example directive that fixes autofocus problem with multiple views
 @Directive({
   selector: '[autofocus]' // using [ ] means selecting attributes
