+++ src/app/pipes/RxPipe.ts
@@ -1,7 +1,6 @@
 /// <reference path="../../typings/_custom.d.ts" />
+import {Pipe, PipeFactory, ObservablePipe} from 'angular2/angular2';
 import {async} from 'angular2/src/change_detection/change_detection';
-import {Pipe, PipeFactory} from 'angular2/change_detection';
-import {ObservablePipe} from 'angular2/pipes';
 import * as Rx from 'rx';
 
 export function isObservable(obs) {
