+++ src/common/shadowDomInjectables.ts
@@ -5,7 +5,7 @@ import {
   EmulatedScopedShadowDomStrategy,
   EmulatedUnscopedShadowDomStrategy
 } from 'angular2/render';
-import {bind} from 'angular2/di';
+import {bind} from 'angular2/angular2';
 import {document} from 'angular2/src/facade/browser';
 import {DOCUMENT_TOKEN} from 'angular2/src/render/dom/dom_renderer';
